+++ src/main/java/de/samply/coder/CoderJob.java
@@ -9,6 +9,7 @@ import de.samply.project.ProjectType;
 import de.samply.project.state.ProjectBridgeheadState;
 import de.samply.query.QueryState;
 import de.samply.register.AppRegisterService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
@@ -16,6 +17,7 @@ import reactor.core.publisher.Mono;
 
 import java.util.List;
 
+@Slf4j
 @Component
 public class CoderJob {
 
@@ -40,13 +42,16 @@ public class CoderJob {
 
     @Scheduled(cron = ProjectManagerConst.CODER_CRON_EXPRESSION_SV)
     public void manageCoderWorkspaces() {
+        log.debug("Starting Coder Job...");
         Mono.when(
                 manageCoderActiveUsers(),
                 manageCoderInactiveUsers()
         ).block();
+        log.debug("Coder Job finished.");
     }
 
     public Mono<Void> manageCoderActiveUsers() {
+        log.debug("Managing Coder Active Users...");
         return Flux.fromIterable(fetchActiveUsers())
                 .flatMap(coderService::createWorkspace)
                 .flatMap(projectCoder -> appRegisterService.register(projectCoder).then(Mono.just(projectCoder)))
@@ -59,6 +64,7 @@ public class CoderJob {
     }
 
     public Mono<Void> manageCoderInactiveUsers() {
+        log.debug("Managing Coder Inactive Users...");
         return Flux.fromIterable(fetchInactiveUsers())
                 .flatMap(coderService::deleteWorkspace)
                 .flatMap(appRegisterService::unregister)